Basic flowfield completed! Yay!

+++ Perlin Noise Flowfield/Particle.cpp
@@ -15,8 +15,8 @@ void Particle::Update(std::vector<std::vector<Vector2>> flowfield, float max_spe
 
 void Particle::Follow(std::vector<std::vector<Vector2>> flowfield, int SCALE)
 {
-    float x = floor(pos.x / SCALE);
-    float y = floor(pos.y / SCALE);
+    float x = floorf(pos.x / SCALE);
+    float y = floorf(pos.y / SCALE);
     Vector2 force = flowfield[x][y];
     ApplyForce(force);
 }
@@ -28,10 +28,10 @@ void Particle::ApplyForce(Vector2 force)
 
 void Particle::WrapAroundEdges(int WIDTH, int HEIGHT)
 {
-    if (pos.x > WIDTH) pos.x = 0;
-    if (pos.x < 0) pos.x = WIDTH;
-    if (pos.y > HEIGHT) pos.y = 0;
-    if (pos.y < 0) pos.y = HEIGHT;
+    if (pos.x > (float)WIDTH) pos.x = 0;
+    if (pos.x < 0) pos.x = (float)WIDTH;
+    if (pos.y > (float)HEIGHT) pos.y = 0;
+    if (pos.y < 0) pos.y = (float)HEIGHT;
 }
 
 void Particle::Draw(float size, Color color)
